remove redundant dtos of user db
