--- src/main/java/ru/yandex/practicum/filmorate/dal/mapper/FriendshipRowMapper.java
@@ -1,19 +1,0 @@
-package ru.yandex.practicum.filmorate.dal.mapper;
-
-import org.springframework.jdbc.core.RowMapper;
-import org.springframework.stereotype.Component;
-import ru.yandex.practicum.filmorate.model.Friendship;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-@Component
-public class FriendshipRowMapper implements RowMapper<Friendship> {
-    @Override
-    public Friendship mapRow(ResultSet rs, int rowNum) throws SQLException {
-        return new Friendship(
-                rs.getLong("user_id"),
-                rs.getLong("friend_id")
-        );
-    }
-}
