+++ src/main/java/ru/yandex/practicum/filmorate/dal/storage/user/UserDbStorage.java
@@ -4,7 +4,6 @@ import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.stereotype.Repository;
 import ru.yandex.practicum.filmorate.dal.storage.BaseDbStorage;
-import ru.yandex.practicum.filmorate.dal.storage.friendship.FriendshipDbStorage;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.util.List;
@@ -18,11 +17,12 @@ public class UserDbStorage extends BaseDbStorage<User> implements UserStorage {
     private static final String FIND_ALL_USERS_QUERY = "SELECT * FROM users";
     private static final String FIND_USER_BY_ID_QUERY = "SELECT * FROM users WHERE id = ?";
 
-    private final FriendshipDbStorage friendshipDbStorage;
+    private static final String INSERT_FRIEND_QUERY = "INSERT INTO friendship (user_id, friend_id) VALUES (?, ?)";
+    private static final String DELETE_FRIEND_QUERY = "DELETE FROM friendship WHERE user_id = ? AND friend_id = ?";
+    private static final String FIND_ALL_BY_USER_ID_QUERY = "SELECT friend_id FROM friendship WHERE user_id = ?";
 
-    public UserDbStorage(JdbcTemplate jdbc, RowMapper<User> mapper, FriendshipDbStorage friendshipDbStorage) {
+    public UserDbStorage(JdbcTemplate jdbc, RowMapper<User> mapper) {
         super(jdbc, mapper);
-        this.friendshipDbStorage = friendshipDbStorage;
     }
 
     @Override
@@ -68,7 +68,7 @@ public class UserDbStorage extends BaseDbStorage<User> implements UserStorage {
 
     @Override
     public List<User> getCommonFriends(long userId, long otherId) {
-        List<Long> commonFriendsIds = friendshipDbStorage.getCommonFriendsIds(userId, otherId);
+        List<Long> commonFriendsIds = getCommonFriendsIds(userId, otherId);
 
         return commonFriendsIds.stream()
                 .map(this::getById)
@@ -79,12 +79,37 @@ public class UserDbStorage extends BaseDbStorage<User> implements UserStorage {
 
     @Override
     public List<User> getFriends(long userId) {
-        List<Long> friendsIds = friendshipDbStorage.getFriendsIds(userId);
+        List<Long> friendsIds = getFriendsIds(userId);
 
         return friendsIds.stream()
                 .map(this::getById)
                 .filter(Optional::isPresent)
                 .map(Optional::get)
                 .toList();
+    }
+
+    @Override
+    public void addFriend(long userId, long friendId) {
+        insert(INSERT_FRIEND_QUERY, userId, friendId);
+    }
+
+    @Override
+    public boolean deleteFriend(long userId, long friendId) {
+        return delete(DELETE_FRIEND_QUERY, userId, friendId);
+    }
+
+    @Override
+    public List<Long> getCommonFriendsIds(long userId, long otherId) {
+        List<Long> userFriendsIds = getFriendsIds(userId);
+        List<Long> otherFriendsIds = getFriendsIds(otherId);
+
+        return userFriendsIds.stream()
+                .filter(otherFriendsIds::contains)
+                .toList();
+    }
+
+    @Override
+    public List<Long> getFriendsIds(long userId) {
+        return jdbc.queryForList(FIND_ALL_BY_USER_ID_QUERY, Long.class, userId);
     }
 }
