+++ src/main/java/ru/yandex/practicum/filmorate/dal/storage/user/UserStorage.java
@@ -9,4 +9,12 @@ public interface UserStorage extends Storage<User> {
     List<User> getFriends(long userId);
 
     List<User> getCommonFriends(long userId, long friendId);
+
+    void addFriend(long userId, long friendId);
+
+    boolean deleteFriend(long userId, long friendId);
+
+    List<Long> getFriendsIds(long userId);
+
+    List<Long> getCommonFriendsIds(long userId, long friendId);
 }
