+++ src/test/java/ru/yandex/practicum/filmorate/dal/storage/user/UserDbStorageTest.java
@@ -6,9 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
-import ru.yandex.practicum.filmorate.dal.mapper.FriendshipRowMapper;
 import ru.yandex.practicum.filmorate.dal.mapper.UserRowMapper;
-import ru.yandex.practicum.filmorate.dal.storage.friendship.FriendshipDbStorage;
 import ru.yandex.practicum.filmorate.model.User;
 
 import java.time.LocalDate;
@@ -21,7 +19,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 @JdbcTest
 @AutoConfigureTestDatabase
 @RequiredArgsConstructor(onConstructor_ = @Autowired)
-@Import({UserDbStorage.class, UserRowMapper.class, FriendshipDbStorage.class, FriendshipRowMapper.class})
+@Import({UserDbStorage.class, UserRowMapper.class})
 class UserDbStorageTest {
     private final UserDbStorage userStorage;
 
@@ -117,5 +115,39 @@ class UserDbStorageTest {
         assertThat(friends)
                 .extracting("id")
                 .containsExactlyInAnyOrder(3L, 5L);
+    }
+
+    @Test
+    void addFriend() {
+        userStorage.addFriend(3, 2);
+
+        List<Long> friendshipIds = userStorage.getFriendsIds(3);
+
+        assertThat(friendshipIds).containsExactlyInAnyOrder(2L);
+    }
+
+    @Test
+    void deleteFriend() {
+        boolean deleted = userStorage.deleteFriend(1, 5);
+
+        assertThat(deleted).isTrue();
+
+        List<Long> friendshipIds = userStorage.getFriendsIds(1);
+
+        assertThat(friendshipIds).containsExactlyInAnyOrder(3L);
+    }
+
+    @Test
+    void getCommonFriendsIds() {
+        List<Long> commonFriendsIds = userStorage.getCommonFriendsIds(1, 2);
+
+        assertThat(commonFriendsIds).containsExactlyInAnyOrder(3L);
+    }
+
+    @Test
+    void getFriendsIds() {
+        List<Long> friendsIds = userStorage.getFriendsIds(1);
+
+        assertThat(friendsIds).containsExactlyInAnyOrder(3L, 5L);
     }
 }
